wip: more changes
