+++ backend/web/emails/hopp-subscription.html
@@ -110,7 +110,7 @@
                 <tr>
                   <td>
                     <a
-                      href="github.com/gethopp/hopp"
+                      href="https://github.com/gethopp/hopp"
                       style="
                         font-size: 0.75rem;
                         line-height: 1rem;
