+++ web-app/src/pages/Onboarding.tsx
@@ -1,34 +1,38 @@
-import { useEffect, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useForm } from "@tanstack/react-form";
 import { useAPI } from "@/hooks/useQueryClients";
 import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
 import { toast } from "react-hot-toast";
 import { CreditCard, ShieldCheck } from "lucide-react";
 import Logo from "@/assets/Hopp.png";
-import { HiMiniCheck } from "react-icons/hi2";
+import { HiMiniCheck, HiOutlineQuestionMarkCircle } from "react-icons/hi2";
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
 
-import { CompanySizeSelect, PairingToolMultiSelect, ReferralSourceSelect } from "@/components/OnboardingModal";
+import { PairingToolMultiSelect, ReferralSourceSelect } from "@/components/onboarding-fields";
 
 const TRIAL_DAYS = 14;
 
-const ONBOARDING_DRAFT_KEY = "hopp:onboarding-draft";
+// Drafts are scoped per account so one admin's onboarding answers can't leak
+// into another account signed in on the same browser.
+const ONBOARDING_DRAFT_PREFIX = "hopp:onboarding-draft";
+
+function draftKey(userId: string): string {
+  return `${ONBOARDING_DRAFT_PREFIX}:${userId}`;
+}
 
 interface OnboardingFormValues {
   teamName: string;
-  companySize: string;
   pairingTool: string[];
-  signUpReason: string;
   hearAboutHopp: string;
   hearAboutHoppOther: string;
 }
 
-function loadDraft(): Partial<OnboardingFormValues> {
+function loadDraft(userId: string): Partial<OnboardingFormValues> {
   try {
-    const stored = localStorage.getItem(ONBOARDING_DRAFT_KEY);
+    const stored = localStorage.getItem(draftKey(userId));
     if (!stored) return {};
     const parsed = JSON.parse(stored);
     return typeof parsed === "object" && parsed !== null ? (parsed as Partial<OnboardingFormValues>) : {};
@@ -37,8 +41,8 @@ function loadDraft(): Partial<OnboardingFormValues> {
   }
 }
 
-function clearDraft() {
-  localStorage.removeItem(ONBOARDING_DRAFT_KEY);
+function clearDraft(userId: string) {
+  localStorage.removeItem(draftKey(userId));
 }
 
 export function Onboarding() {
@@ -54,12 +58,16 @@ export function Onboarding() {
   const subscription = subscriptionData?.subscription;
   const needsOnboarding = !!subscription?.requires_payment_method && !!subscription?.is_admin;
 
+  // The current account id scopes the localStorage draft to this user.
+  const { data: currentUser } = useQuery("get", "/api/auth/user", undefined, {});
+  const userId = currentUser?.id;
+
   useEffect(() => {
     if (subscription && !needsOnboarding) {
-      clearDraft();
+      if (userId) clearDraft(userId);
       navigate("/dashboard", { replace: true });
     }
-  }, [subscription, needsOnboarding, navigate]);
+  }, [subscription, needsOnboarding, navigate, userId]);
 
   const { mutateAsync: updateTeam } = useMutation("patch", "/api/auth/team");
   const { mutateAsync: updateOnboardingFormStatus } = useMutation("post", "/api/auth/metadata/onboarding-form");
@@ -68,18 +76,14 @@ export function Onboarding() {
   const form = useForm({
     defaultValues: {
       teamName: "",
-      companySize: "",
       pairingTool: [] as string[],
-      signUpReason: "",
       hearAboutHopp: "",
       hearAboutHoppOther: "",
-      ...loadDraft(),
     },
     onSubmit: async ({ value }: { value: OnboardingFormValues }) => {
       const teamName = value.teamName.trim();
       if (
         teamName === "" ||
-        value.companySize === "" ||
         value.pairingTool.length === 0 ||
         value.hearAboutHopp === "" ||
         (value.hearAboutHopp === "other" && value.hearAboutHoppOther === "")
@@ -94,9 +98,7 @@ export function Onboarding() {
         await updateOnboardingFormStatus({
           body: {
             onboarding: {
-              companySize: value.companySize,
               pairingTool: value.pairingTool,
-              signUpReason: value.signUpReason,
               hearAboutHopp: value.hearAboutHopp,
               hearAboutHoppOther: value.hearAboutHoppOther,
             },
@@ -110,12 +112,25 @@ export function Onboarding() {
     },
   });
 
+  // Load the per-account draft once the current user is known. Deferred (instead
+  // of in defaultValues) because the user id resolves asynchronously.
+  const draftLoadedRef = useRef(false);
   useEffect(() => {
+    if (!userId || draftLoadedRef.current) return;
+    draftLoadedRef.current = true;
+    const draft = loadDraft(userId);
+    if (Object.keys(draft).length > 0) {
+      form.reset({ ...form.state.values, ...draft });
+    }
+  }, [userId, form]);
+
+  useEffect(() => {
+    if (!userId) return;
     const subscription = form.store.subscribe(() => {
-      localStorage.setItem(ONBOARDING_DRAFT_KEY, JSON.stringify(form.store.state.values));
+      localStorage.setItem(draftKey(userId), JSON.stringify(form.store.state.values));
     });
     return () => subscription.unsubscribe();
-  }, [form]);
+  }, [form, userId]);
 
   const handleStartTrial = async () => {
     setCheckoutLoading(true);
@@ -152,69 +167,66 @@ export function Onboarding() {
 
         <div className="rounded-2xl border border-slate-200 bg-white p-8 shadow-xl ring-2 ring-white">
           {step === 1 ?
-            <form
-              onSubmit={(e) => {
-                e.preventDefault();
-                form.handleSubmit();
-              }}
-              className="space-y-6"
-            >
-              <div>
-                <h1 className="text-2xl font-semibold">Tell us about your team</h1>
-                <p className="mt-2 text-base font-normal text-gray-600">
-                  A few quick questions so we can set up your workspace.
-                </p>
+            <div className="flex flex-col gap-4">
+              <form
+                onSubmit={(e) => {
+                  e.preventDefault();
+                  form.handleSubmit();
+                }}
+                className="space-y-6"
+              >
+                <div>
+                  <h1 className="text-2xl font-semibold">Tell us about your team</h1>
+                  <p className="mt-2 text-base font-normal text-gray-600">
+                    A few quick questions so we can set up your workspace.
+                  </p>
+                </div>
+
+                <form.Field name="teamName">
+                  {(field) => {
+                    const hasError = showValidationErrors && field.state.value.trim() === "";
+                    return (
+                      <div className="space-y-2">
+                        <Label htmlFor="teamName">
+                          Team name <span className="text-red-500">*</span>
+                        </Label>
+                        <Input
+                          id="teamName"
+                          value={field.state.value}
+                          onChange={(e) => field.handleChange(e.target.value)}
+                          placeholder="Acme Engineering"
+                        />
+                        {hasError && <p className="text-sm text-red-500">Team name is required</p>}
+                      </div>
+                    );
+                  }}
+                </form.Field>
+
+                <ReferralSourceSelect form={form} showErrors={showValidationErrors} />
+                <PairingToolMultiSelect form={form} showErrors={showValidationErrors} />
+
+                <form.Subscribe selector={(state) => ({ isSubmitting: state.isSubmitting })}>
+                  {({ isSubmitting }: { isSubmitting: boolean }) => (
+                    <div className="flex justify-end pt-2">
+                      <Button type="submit" disabled={isSubmitting}>
+                        {isSubmitting ? "Saving..." : "Continue"}
+                      </Button>
+                    </div>
+                  )}
+                </form.Subscribe>
+              </form>
+
+              <div className="flex items-center gap-4 mx-6">
+                <span className="h-px flex-1 bg-slate-200" />
+                <span className="text-xs font-medium text-slate-400">Or</span>
+                <span className="h-px flex-1 bg-slate-200" />
               </div>
 
-              <form.Field name="teamName">
-                {(field) => {
-                  const hasError = showValidationErrors && field.state.value.trim() === "";
-                  return (
-                    <div className="space-y-2">
-                      <Label htmlFor="teamName">
-                        Team name <span className="text-red-500">*</span>
-                      </Label>
-                      <Input
-                        id="teamName"
-                        value={field.state.value}
-                        onChange={(e) => field.handleChange(e.target.value)}
-                        placeholder="Acme Engineering"
-                      />
-                      {hasError && <p className="text-sm text-red-500">Team name is required</p>}
-                    </div>
-                  );
-                }}
-              </form.Field>
-
-              <ReferralSourceSelect form={form} showErrors={showValidationErrors} />
-              <CompanySizeSelect form={form} showErrors={showValidationErrors} />
-              <PairingToolMultiSelect form={form} showErrors={showValidationErrors} />
-
-              <form.Field name="signUpReason">
-                {(field) => (
-                  <div className="space-y-2">
-                    <Label htmlFor="signUpReason">Why did you sign up for Hopp?</Label>
-                    <Textarea
-                      id="signUpReason"
-                      value={field.state.value}
-                      onChange={(e) => field.handleChange(e.target.value)}
-                      placeholder="Tell us why you signed up..."
-                      rows={3}
-                    />
-                  </div>
-                )}
-              </form.Field>
-
-              <form.Subscribe selector={(state) => ({ isSubmitting: state.isSubmitting })}>
-                {({ isSubmitting }: { isSubmitting: boolean }) => (
-                  <div className="flex justify-end pt-2">
-                    <Button type="submit" disabled={isSubmitting}>
-                      {isSubmitting ? "Saving..." : "Continue"}
-                    </Button>
-                  </div>
-                )}
-              </form.Subscribe>
-            </form>
+              <div className="space-y-1 ml-6">
+                <h3 className="text-md font-semibold">Want to join an existing team?</h3>
+                <p className="text-sm text-muted-foreground">You'll need an invite link from a teammate.</p>
+              </div>
+            </div>
           : <div className="space-y-6">
               <div>
                 <h1 className="text-2xl font-semibold">Start your {TRIAL_DAYS}-day free trial</h1>
@@ -222,6 +234,29 @@ export function Onboarding() {
                   Add a payment method to unlock Hopp. You won't be charged today and your card is only billed when the{" "}
                   {TRIAL_DAYS}-day trial ends, and you can cancel anytime before then with one click.
                 </p>
+                <TooltipProvider delayDuration={0.2}>
+                  <Tooltip>
+                    <TooltipTrigger asChild>
+                      <button
+                        type="button"
+                        className="mt-3 inline-flex items-center gap-1.5 text-sm font-normal text-indigo-600 hover:text-indigo-700"
+                      >
+                        <HiOutlineQuestionMarkCircle className="size-5 stroke-2" />
+                        Why do you require a credit card?
+                      </button>
+                    </TooltipTrigger>
+                    <TooltipContent className="max-w-xs space-y-2 text-left leading-relaxed">
+                      <p>
+                        We require a credit card to avoid the noise of people who don't want to commit using the hosted
+                        Hopp solution.
+                      </p>
+                      <p className="mb-0">
+                        We have costs for media sharing, and we work on this nights and weekends, so we want to spend
+                        all our bandwidth on teams that care most about pairing.
+                      </p>
+                    </TooltipContent>
+                  </Tooltip>
+                </TooltipProvider>
               </div>
 
               <ul className="space-y-3 text-sm text-gray-700">
